third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/origin-whitelisting-https-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt?sync. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt?async. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Tests that origin whitelisting for https does not match http URLs.

Testing: http://localhost:8000/xmlhttprequest/resources/get.txt (sync)
PASS: NetworkError: A network error occurred.
Testing: http://localhost:8000/xmlhttprequest/resources/get.txt (async)
PASS: 0
